move spline typing to ffd - avoids cylic import
